// src/domain/models/enums/permission.enum.ts
export enum PermissionKey {
  UserRead = 'USER:READ',
  UserCreate = 'USER:CREATE',
  UserUpdate = 'USER:UPDATE',
  UserDelete = 'USER:DELETE',

  RoleRead = 'ROLE:READ',
  RoleCreate = 'ROLE:CREATE',
  RoleUpdate = 'ROLE:UPDATE',
  RoleDelete = 'ROLE:DELETE',

  PermissionRead = 'PERMISSION:READ',
  PermissionCreate = 'PERMISSION:CREATE',
  PermissionUpdate = 'PERMISSION:UPDATE',
  PermissionDelete = 'PERMISSION:DELETE',
}